Fixed Graph.js
Fixes bug caused by calculating missingUnits same way whether just starting loop or in the middle of it.

When just starting loop currentUnit - startUnit is the number of null entries needed whereas after first input it is currentUnit - startUnit - 1.

Having said that this may actually still be a bug. Keep an eye on it.